refactor: :bulb: split logic codes
